KEYCLOAK-11069 auth-server-remote tests
