+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/arquillian/containers/KeycloakContainerEventsController.java
@@ -18,15 +18,9 @@ package org.keycloak.testsuite.arquillian.containers;
 
 import org.apache.commons.io.FileUtils;
 import org.jboss.arquillian.config.descriptor.api.ContainerDef;
-import org.jboss.arquillian.container.impl.client.container.ContainerDeployController;
 import org.jboss.arquillian.container.spi.Container;
 import org.jboss.arquillian.container.spi.ContainerRegistry;
-import org.jboss.arquillian.container.spi.client.deployment.Deployment;
-import org.jboss.arquillian.container.spi.client.deployment.DeploymentScenario;
 import org.jboss.arquillian.container.spi.event.ContainerMultiControlEvent;
-import org.jboss.arquillian.container.spi.event.DeployDeployment;
-import org.jboss.arquillian.container.spi.event.DeployManagedDeployments;
-import org.jboss.arquillian.container.spi.event.DeploymentEvent;
 import org.jboss.arquillian.container.spi.event.StartClassContainers;
 import org.jboss.arquillian.container.spi.event.StartSuiteContainers;
 import org.jboss.arquillian.container.spi.event.StopClassContainers;
@@ -35,7 +29,6 @@ import org.jboss.arquillian.container.spi.event.StopSuiteContainers;
 import org.jboss.arquillian.container.spi.event.UnDeployManagedDeployments;
 import org.jboss.arquillian.container.test.impl.client.ContainerEventController;
 import org.jboss.arquillian.core.api.Event;
-import org.jboss.arquillian.core.api.Injector;
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.core.api.annotation.Observes;
@@ -49,7 +42,6 @@ import org.jboss.shrinkwrap.api.exporter.ZipExporter;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.keycloak.common.Profile;
 import org.keycloak.helpers.DropAllServlet;
-import org.keycloak.testsuite.arquillian.AuthServerTestEnricher;
 import org.keycloak.testsuite.arquillian.ContainerInfo;
 import org.keycloak.testsuite.arquillian.annotation.RestartContainer;
 import org.wildfly.extras.creaper.commands.deployments.Deploy;
@@ -67,9 +59,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Optional;
 import java.util.Properties;
+import org.keycloak.testsuite.util.ContainerAssume;
 
 /**
  * Changes behaviour of original ContainerEventController to stop manual containers 
@@ -87,6 +79,8 @@ public class KeycloakContainerEventsController extends ContainerEventController 
 
     @Inject
     private Event<ContainerMultiControlEvent> container;
+    @Inject
+    private Instance<ContainerRegistry> containerRegistry;
 
     @Override
     public void execute(@Observes AfterSuite event) {
@@ -109,6 +103,10 @@ public class KeycloakContainerEventsController extends ContainerEventController 
     @Override
     public void execute(BeforeClass event) {
         if (event.getTestClass().isAnnotationPresent(RestartContainer.class)) {
+
+            // stop executing the test - remote container cannot be restarted
+            ContainerAssume.assumeNotAuthServerRemote();
+
             RestartContainer restartContainer = event.getTestClass().getAnnotation(RestartContainer.class);
 
             beforeOriginalContainerStop(restartContainer);
@@ -340,54 +338,6 @@ public class KeycloakContainerEventsController extends ContainerEventController 
             } catch (IOException e) {
                 e.printStackTrace();
             }
-        }
-    }
-
-    /*
-     * Coppied from org.jboss.arquillian.container.impl.client.container.ContainerDeployController
-     * 
-     * Overrides a condition that container cannot be in manual mode, and deploys the deployment
-     * if the container is started
-     */
-    @Inject
-    private Instance<Injector> injector;
-    @Inject
-    private Instance<DeploymentScenario> deploymentScenario;
-    @Inject
-    private Instance<ContainerRegistry> containerRegistry;
-
-    public void deployManaged(@Observes DeployManagedDeployments event) throws Exception {
-        forEachManagedDeployment(new ContainerDeployController.Operation<Container, Deployment>() {
-            @Inject
-            private Event<DeploymentEvent> event;
-
-            @Override
-            public void perform(Container container, Deployment deployment) throws Exception {
-                if (container.getState().equals(Container.State.STARTED)) {
-                    event.fire(new DeployDeployment(container, deployment));
-                }
-            }
-        });
-    }
-
-    private void forEachManagedDeployment(ContainerDeployController.Operation<Container, Deployment> operation) throws Exception {
-        DeploymentScenario scenario = this.deploymentScenario.get();
-        if (scenario == null) {
-            return;
-        }
-        forEachDeployment(scenario.managedDeploymentsInDeployOrder(), operation);
-    }
-
-    private void forEachDeployment(List<Deployment> deployments, ContainerDeployController.Operation<Container, Deployment> operation)
-        throws Exception {
-        injector.get().inject(operation);
-        ContainerRegistry containerRegistry = this.containerRegistry.get();
-        if (containerRegistry == null) {
-            return;
-        }
-        for (Deployment deployment : deployments) {
-            Container container = containerRegistry.getContainer(deployment.getDescription().getTarget());
-            operation.perform(container, deployment);
         }
     }
 }
