+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/util/ContainerAssume.java
@@ -35,6 +35,11 @@ public class ContainerAssume {
     }
     
 
+    public static void assumeNotAuthServerRemote() {
+        Assume.assumeFalse("Doesn't work on auth-server-remote", 
+                AuthServerTestEnricher.AUTH_SERVER_CONTAINER.equals("auth-server-remote"));
+    }
+
     public static void assumeClusteredContainer() {
         Assume.assumeTrue(
               String.format("Ignoring test since %s is set to false",
