+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/model/ConcurrentTransactionsTest.java
@@ -35,6 +35,9 @@ import java.util.Arrays;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
 
 import static org.keycloak.testsuite.arquillian.DeploymentTargetModifier.AUTH_SERVER_CURRENT;
 
@@ -284,7 +287,8 @@ public class ConcurrentTransactionsTest extends AbstractTestRealmKeycloakTest {
             um.removeUser(realm, realmUser2);
         }
 
-        assert (currentSession.realms().removeRealm(realm.getId()));
+        Assert.assertTrue(currentSession.realms().removeRealm(realm.getId()));
+        Assert.assertThat(currentSession.realms().getRealm(realm.getId()), is(nullValue()));
     }
 
     @Override
